Rename : Layout을 components 폴더로 이동
- Layout을 components 폴더로 이동하고 router.jsx의 import 경로 수정
