+++ src/router/router.jsx
@@ -1,6 +1,6 @@
 import { createBrowserRouter } from 'react-router-dom';
 import Home from '../pages/Home';
-import Layout from '../pages/Layout';
+import Layout from '../components/Layout';
 import List from '../pages/List/List';
 import Post from '../pages/Post';
